Use render function as system
